--- a/app/src/router/index.ts
+++ b/app/src/router/index.ts
@@ -1,6 +1,7 @@
 import { ad4mConnect } from "@/ad4mConnect";
 import { useAppStore, useRouteMemoryStore } from "@/stores";
 import { RouteParams } from "@coasys/flux-types";
+import { storeToRefs } from "pinia";
 import { createRouter, createWebHashHistory, RouteRecordRaw } from "vue-router";
 
 const routes: Array<RouteRecordRaw> = [
@@ -69,7 +70,11 @@ router.beforeEach(async (to, from, next) => {
   try {
     const isAuthenticated = await ad4mConnect.isAuthenticated();
     if (isAuthenticated) {
-      const { me } = useAppStore();
+      const appStore = useAppStore();
+      const { ad4mClient } = storeToRefs(appStore);
+      const me = await ad4mClient.value.agent.me();
+
+      // Handle authenticated routes
       const fluxAccountCreated = me.perspective?.links.find((e) => e.data.source.startsWith("flux://"));
       const isOnSignupOrMain = to.name === "signup" || to.name === "main";
       if (fluxAccountCreated && isOnSignupOrMain) next("/home");