--- a/app/src/router/index.ts
+++ b/app/src/router/index.ts
@@ -71,11 +71,11 @@ router.beforeEach(async (to, from, next) => {
     const isAuthenticated = await ad4mConnect.isAuthenticated();
     if (isAuthenticated) {
       const appStore = useAppStore();
-      const { ad4mClient } = storeToRefs(appStore);
-      const me = await ad4mClient.value.agent.me();
+      await appStore.refreshMyProfile();
+      const { me } = storeToRefs(appStore);
 
       // Handle authenticated routes
-      const fluxAccountCreated = me.perspective?.links.find((e) => e.data.source.startsWith("flux://"));
+      const fluxAccountCreated = me.value.perspective?.links.find((e) => e.data.source.startsWith("flux://"));
       const isOnSignupOrMain = to.name === "signup" || to.name === "main";
       if (fluxAccountCreated && isOnSignupOrMain) next("/home");
       if (!fluxAccountCreated && !isOnSignupOrMain) next("/signup");