--- a/app/src/stores/aiStore.ts
+++ b/app/src/stores/aiStore.ts
@@ -44,6 +44,7 @@ export const useAiStore = defineStore(
     const transcriptionModel = ref("Base");
     const transcriptionPreviewTimeout = ref(0.4);
     const transcriptionMessageTimeout = ref(3);
+    const transcriptionMaxChars = ref(1000);
 
     function setTranscriptionEnabled(payload: boolean): void {
       transcriptionEnabled.value = payload;
@@ -124,6 +125,7 @@ export const useAiStore = defineStore(
       transcriptionModel,
       transcriptionPreviewTimeout,
       transcriptionMessageTimeout,
+      transcriptionMaxChars,
       setTranscriptionEnabled,
       toggleTranscriptionEnabled,
       setTranscriptionModel,