- Image upload fixes
- UI display fix for community list images
- Transcription timeout decreased from 5 to 3 seconds
- 1000 max char limit added to transcriber
- Sign up routing & profile data fixes
- Processing prevented when community is syncing

<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

* **New Features**
  * Added enforcement of a maximum character limit for transcription messages, with automatic saving when the limit is reached.
  * Introduced synchronization state events for communities, enabling external listeners (such as web components) to react to sync status changes.

* **Improvements**
  * Avatar cropping components now use correct capitalization, ensuring proper rendering.
  * Avatar wrapper elements now have consistent sizing in community lists.
  * Transcription message timeout reduced from 5 to 3 seconds for faster saving.
  * User profile management is streamlined and now handled centrally via the app store, with profile refreshes integrated after updates and during navigation.
  * Navigation and event handling in the signup process are simplified and clarified.
  * Profile views now automatically update when the user's profile changes.
  * Cached profile fetching improved with optional refresh capability.

* **Bug Fixes**
  * Timeline columns in the Synergy demo now respect community synchronization state, preventing processing when unsynced.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->